requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
pytesseract==0.3.13
PyMuPDF==1.24.9
google-cloud-documentai>=2.20.0
PyYAML==6.0.2
python-dotenv>=1.0.0
numpy>=1.24.0,<3.0.0
python-docx==1.1.2
joblib>=1.3.0
scikit-learn==1.8.0
scipy>=1.10.0

# RAG (multilingual embeddings) - pinned for reproducibility
sentence-transformers==2.7.0
faiss-cpu==1.7.4
chromadb==0.4.22
# Local LLM (LFM2.5-1.2B-Instruct from disk)
transformers==4.45.0
torch==2.2.0
accelerate==0.33.0

# =========================
# Backend enhancements
# =========================
passlib[bcrypt]==1.7.4   # password hashing
sqlalchemy==2.0.32       # SQLite ORM
slowapi==0.1.9           # rate limiting
structlog==24.4.0        # structured logging
python-jose[cryptography]==3.3.0
email-validator==2.2.0
httpx>=0.27.0  # Google OAuth token exchange
google-genai>=1.0.0  # Gemini chat (current SDK; use GEMINI_MODEL=gemini-2.0-flash)